backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-dotenv
pymongo

langchain
langchain-community
langchain-google-genai
google-generativeai